Refactor: Move QS logic from reader post list vm to parent reader vm
